Drasil.Metadata

Drasil Metadata

drasilMeta :: DrasilMeta

Drasil metadata.

watermark :: String

Drasil's watermark for generated files.

Domains

compScience :: IdeaDict

For ideas, concepts, or terms related to Computer Science.

Various domains that are used in Drasil. May have an abbreviation.

softEng :: IdeaDict

For ideas, concepts, or terms related to Software Engineering.

Various domains that are used in Drasil. May have an abbreviation.

mathematics :: IdeaDict

For ideas, concepts, or terms related to Mathematics.

Various domains that are used in Drasil. May have an abbreviation.

progLanguage :: IdeaDict

For ideas, concepts, or terms related to Programming Languages.

Various domains that are used in Drasil. May have an abbreviation.

physics :: IdeaDict

For ideas, concepts, or terms related to Physics.

Various domains that are used in Drasil. May have an abbreviation.

civilEng :: IdeaDict

For ideas, concepts, or terms related to Civil Engineering.

Various domains that are used in Drasil. May have an abbreviation.

materialEng :: IdeaDict

For ideas, concepts, or terms related to Material Engineering.

Various domains that are used in Drasil. May have an abbreviation.

documentc :: IdeaDict

For ideas, concepts, or terms related to Documents.

Various domains that are used in Drasil. May have an abbreviation.

knowledgemng :: IdeaDict

For ideas, concepts, or terms related to Knowledge Management.

Various domains that are used in Drasil. May have an abbreviation.

Supported Software

Supported Software

runnableSoftware :: IdeaDict

website :: IdeaDict

Theory Concepts

dataDefn :: CI

Data definition.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.

genDefn :: CI

General definition.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.

inModel :: CI

Instance model.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.

thModel :: CI

Theoretical model.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.