Drasil.Metadata.TheoryConcepts

Description

Theory related Drasil concepts, used across Drasil.

Documentation

dataDefn :: CI

Data definition.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.

genDefn :: CI

General definition.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.

inModel :: CI

Instance model.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.

thModel :: CI

Theoretical model.

These are internal-to-Drasil common ideas, and need to be defined at the same time as theories.