Drasil.Metadata.TraceabilityGraphs

Documentation

data GraphInfo

Holds all important and relevant information for generating a traceability graph. Includes nodes, graph edges, and node family information.

Constructors

GI

Fields

assumpNF :: NodeFamily

Assumptions.

ddNF :: NodeFamily

Data definitions.

gdNF :: NodeFamily

General definitions.

tmNF :: NodeFamily

Theory models.

imNF :: NodeFamily

Instance models.

reqNF :: NodeFamily

Requirements (both functional and non-functional).

gsNF :: NodeFamily

Goal statement.

chgNF :: NodeFamily

Changes (both likely and unlikely).

edgesAvsA :: [(UID, [UID])]

Assumptions dependent on assumptions.

edgesAvsAll :: [(UID, [UID])]

Definitions, models, requirements, and changes dependent on assumptions.

edgesRefvsRef :: [(UID, [UID])]

Definitions and models that are dependent on other definitions and models.

edgesAllvsR :: [(UID, [UID])]

Goals and requirements dependent on definitions, models, and other requirements.

edgesAllvsAll :: [(UID, [UID])]

Definitions, models, requirements, goals, and changes that are dependent on one another.

data NodeFamily

A node family contains a list of UIDs, their display labels, general subgraph label, and colour.

Constructors

NF

Fields

nodeUIDs :: [UID]

Node UIDs.

nodeLabels :: [Label]

Display labels for nodes. We use the reference addresses from the UIDs.

nfLabel :: Label

Individual subgraph labels. These labels do not show on the final generated pdf or png files.

nfColour :: Colour

Gives the ability to change colours of bubbles within the graph.

type Label = String

Type synonym for clarity.

type Colour = String

Type synonym for clarity.